Tweaked ft_itoa_base to accommodate longer values

+++ libft/ft_itoa_base.c
@@ -6,27 +6,27 @@
 /*   By: tlouekar <tlouekar@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/21 14:02:58 by exam              #+#    #+#             */
-/*   Updated: 2020/01/27 11:41:26 by tlouekar         ###   ########.fr       */
+/*   Updated: 2020/01/28 10:29:15 by tlouekar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/libft.h"
 
-char		*ft_itoa_base(int value, int base)
+char		*ft_itoa_base(long value, int base)
 {
 	char	*s;
 	long	n;
 	int		sign;
 	int		i;
 
-	n = (value < 0) ? -(long)value : value;
+	n = (value < 0) ? -value : value;
 	sign = (value < 0 && base == 10) ? -1 : 0;
 	i = (sign == -1) ? 2 : 1;
 	while ((n /= base) >= 1)
 		i++;
 	s = (char*)malloc(sizeof(char) * (i + 1));
 	s[i] = '\0';
-	n = (value < 0) ? -(long)value : value;
+	n = (value < 0) ? -value : value;
 	while (i-- + sign)
 	{
 		s[i] = (n % base < 10) ? n % base + '0' : n % base + 'A' - 10;
